Added custom hook for determining when update animation should show.
