+++ assets/css/animation.scss
@@ -1,4 +1,4 @@
-.animation {
+.update-animation {
   background-color: rgba(0, 117, 219, 0.24);
   mix-blend-mode: normal;
   position: absolute;
