+++ assets/js/components/Dashboard/PlacesPage.tsx
@@ -1,4 +1,4 @@
-import React, { ComponentType, useEffect, useState } from "react";
+import React, { ComponentType } from "react";
 import PlaceRow from "./PlaceRow";
 import PlacesActionBar from "./PlacesActionBar";
 import FilterDropdown from "./FilterDropdown";
@@ -70,21 +70,7 @@ const PlacesList: ComponentType<PlacesListProps> = ({
     activeEventKeys,
   } = usePlacesPageContext();
   const dispatch = usePlacesPageDispatchContext();
-  const prevPlaces = usePrevious(places);
-  const [newPlaceIds, setNewPlaceIds] = useState<string[]>([]);
-
-  useEffect(() => {
-    if (prevPlaces) {
-      const placeIds = places
-        .filter((place) => !prevPlaces.includes(place))
-        .map((place) => place.id);
-
-      setNewPlaceIds(placeIds);
-      const timer = setTimeout(() => setNewPlaceIds([]), 2000);
-
-      return () => clearTimeout(timer);
-    }
-  }, [JSON.stringify(places)]);
+  const prevPlaceIds = usePrevious(places)?.map((place) => place.id);
 
   const handleClickResetFilters = () => {
     dispatch({ type: "RESET_STATE", page: "PLACES" });
@@ -316,7 +302,7 @@ const PlacesList: ComponentType<PlacesListProps> = ({
               classNames={isFiltered || isAlertPlacesList ? "filtered" : ""}
               filteredLine={isOnlyFilteredByRoute ? getFilteredLine() : null}
               defaultSort={sortDirection === 0}
-              showAnimation={newPlaceIds.includes(place.id)}
+              showAnimation={!prevPlaceIds.includes(place.id)}
             />
           );
         })}
